srcs/backend/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-rest-framework==0.1.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.1
sqlparse==0.5.1
djangorestframework-simplejwt==5.3.1
django-oauth-toolkit==3.0.1